backend/requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
mypy-extensions==1.0.0
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
typing-inspect==0.9.0
typing_extensions==4.8.0
Werkzeug==3.0.0
python-dotenv==1.0.0
